Renamed test_data to test_providers
