+++ tests/test_providers/test_hardware.py
